Refactor SMM initialization.
